+++ daemon/src/ctrl_power.rs
@@ -12,6 +12,7 @@ use zbus::export::futures_util::lock::Mutex;
 use zbus::{dbus_interface, Connection, SignalContext};
 
 use crate::config::Config;
+use crate::config_traits::StdConfig;
 use crate::error::RogError;
 use crate::{task_watch_item, CtrlTask, GetSupported};