+++ daemon/src/ctrl_profiles/controller.rs
@@ -5,11 +5,12 @@ use rog_profiles::error::ProfileError;
 use rog_profiles::{FanCurveProfiles, Profile};
 
 use super::config::ProfileConfig;
+use crate::config_traits::StdConfig;
 use crate::error::RogError;
 use crate::GetSupported;
 
 pub struct CtrlPlatformProfile {
-    pub config: ProfileConfig,
+    pub profile_config: ProfileConfig,
     pub platform: AsusPlatform,
 }
 
@@ -50,17 +51,17 @@ impl CtrlPlatformProfile {
         if platform.has_platform_profile() || platform.has_throttle_thermal_policy() {
             info!("Device has profile control available");
 
-            let mut controller = CtrlPlatformProfile { config, platform };
+            let mut controller = CtrlPlatformProfile { profile_config: config, platform };
             if FanCurveProfiles::get_device().is_ok() {
                 info!("Device has fan curves available");
-                if controller.config.fan_curves.is_none() {
-                    controller.config.fan_curves = Some(Default::default());
+                if controller.profile_config.fan_curves.is_none() {
+                    controller.profile_config.fan_curves = Some(Default::default());
                     for _ in [Profile::Balanced, Profile::Performance, Profile::Quiet] {
                         controller.set_next_profile()?;
                         controller.set_active_curve_to_defaults()?;
 
                         let active = Profile::get_active_profile().unwrap_or(Profile::Balanced);
-                        if let Some(curves) = controller.config.fan_curves.as_ref() {
+                        if let Some(curves) = controller.profile_config.fan_curves.as_ref() {
                             info!(
                                 "{active:?}: {}",
                                 String::from(curves.get_fan_curves_for(active))
@@ -77,25 +78,25 @@ impl CtrlPlatformProfile {
     }
 
     pub fn save_config(&self) {
-        self.config.write();
+        self.profile_config.write();
     }
 
     /// Toggle to next profile in list. This will first read the config, switch,
     /// then write out
     pub(super) fn set_next_profile(&mut self) -> Result<(), RogError> {
         // Read first just incase the user has modified the config before calling this
-        match self.config.active_profile {
+        match self.profile_config.active_profile {
             Profile::Balanced => {
                 Profile::set_profile(Profile::Performance)?;
-                self.config.active_profile = Profile::Performance;
+                self.profile_config.active_profile = Profile::Performance;
             }
             Profile::Performance => {
                 Profile::set_profile(Profile::Quiet)?;
-                self.config.active_profile = Profile::Quiet;
+                self.profile_config.active_profile = Profile::Quiet;
             }
             Profile::Quiet => {
                 Profile::set_profile(Profile::Balanced)?;
-                self.config.active_profile = Profile::Balanced;
+                self.profile_config.active_profile = Profile::Balanced;
             }
         }
         self.write_profile_curve_to_platform()?;
@@ -104,18 +105,18 @@ impl CtrlPlatformProfile {
 
     /// Set the curve for the active profile active
     pub(super) fn write_profile_curve_to_platform(&mut self) -> Result<(), RogError> {
-        if let Some(curves) = &mut self.config.fan_curves {
+        if let Some(curves) = &mut self.profile_config.fan_curves {
             if let Ok(mut device) = FanCurveProfiles::get_device() {
-                curves.write_profile_curve_to_platform(self.config.active_profile, &mut device)?;
+                curves.write_profile_curve_to_platform(self.profile_config.active_profile, &mut device)?;
             }
         }
         Ok(())
     }
 
     pub(super) fn set_active_curve_to_defaults(&mut self) -> Result<(), RogError> {
-        if let Some(curves) = self.config.fan_curves.as_mut() {
+        if let Some(curves) = self.profile_config.fan_curves.as_mut() {
             if let Ok(mut device) = FanCurveProfiles::get_device() {
-                curves.set_active_curve_to_defaults(self.config.active_profile, &mut device)?;
+                curves.set_active_curve_to_defaults(self.profile_config.active_profile, &mut device)?;
             }
         }
         Ok(())